Update README.md
Robot  is  the  word  i include  it. But Artificial Intelligence (AI) is  based on software. This AI is  included in  my professional subject. So this  lead to confusion that iam saying about AI. This is  the text correction i have made it. To make it  understand for any visitors  visiting  the  website..